Add files via upload
Changed the flights.jsx to add an origin and final destination, as well as choosing date of departure and date of return if applicable. Also catches any errors that may occur with inputs, please note that although stuff does pop up, there is a blank page when providing a valid origin -> destination and departure/return dates. This error seems to continuously occur on line 115 and must be fixed.